protocol: make block committing idempotent
Previously, we only called CommitAppliedBlock from a single goroutine:
either the core/generator or the core/fetch goroutine. This meant we
could access c.State() and apply new blocks to it without the possiblity
of another goroutine applying newer blocks first.

When we modify Chain to always have an up-to-date state tree regardless
of leader status, the goroutine listening to txdb.ListenBlocks heights
will need to also apply blocks. In preparation for this change, we need
to make committing & applying a block idempotent.